+++ src/service/product-validator.js
